client gui begining

+++ server/src/game/Player.kt
@@ -51,7 +51,7 @@ data class Player(val id: Int, val userId: UserId, val state: PlayerState) {
 
         if (direction != null) {
             val angleDiff = state.orientation.orientedAngleWithVector(direction)
-            val maxAngle = (game.properties.playerRotationSpeed / (1000F / updateTime)).toDouble()
+            val maxAngle = (game.properties.playerRotationSpeed / (1000F / updateTime))
             if (abs(angleDiff) <= maxAngle) {
                 state.orientation = state.orientation.rotated(angleDiff)
                 state.orientationTarget = null
